Wedge: kernel module autoload cleanup
Summary:
Since wedge100 and galaxy100 are based on wedge, any kernel module
autoload defined in wedge will be autoloaded in wedge100 and galaxy100.
This causes issue to ast_adc as it is not used in wedge100 and
galaxy100.

As we move to use setup_i2c.sh to add i2c devices now, there is no need
to autoload kernel module.

The diff removes kernel module autoload and directly do modprobe ast_adc
on wedge.

Test Plan:
Load the image to wedge and wedge100, make sure ast_adc is
loaded in wedge and not in wedge100

Reviewed By: jmehta

fbshipit-source-id: d5bba67