vboot: Remove example kek and use DTC_FLAGS
Summary:
This diff includes several changes to vboot that fix the 4:40 error state.

We remove the example KEK public certificates as they are not used.
We move the `DTC_FLAGS` to an environment variable for U-Boot's `make` call.
We increase the reset time for sha256:rsa4096 calculations to 2mins.

Test Plan: When using two signed SPIs this should verify correct with a 0:0 exit code for `vboot-util`.

Reviewed By: sdasari73

fbshipit-source-id: fd2b1e2