vboot: Add TPM, NV, and rollback error text to vboot-util
Summary:
This fills out textual errors for TPM, TPM NV, and rollback protection checks.
The most common error will be `7:70`, where the TPM hardware is in an invalid state.

Test Plan:
Run the `create-tests.sh`, flash CS0 `0.0` to both CS0, CS1.
Check for `0.0`. If `7:70` is output, expect a textual response.

Reviewed By: amithash

fbshipit-source-id: 6e7d523