healthd: Demote CPU utilization messages from critical to warning
Summary:
On certain operations and stress tests it is valid for the CPU utilization to at a high level for some extended period of time.
But we still need to have the messages to detect abnormal cases. Hence demote the criticality of the message from CRIT to WARNING

Test Plan: Test runtime on fbtp.

Reviewed By: sdasari73

fbshipit-source-id: c263b45