fbtp: Refactor VR related functionality into its own library
Summary:
Move all VR related functionality from pal.c/h to vr.c/h.
This at the minimum better structures the code, but the hope is for this to be the first step in
trying to recognize if VR related functionality can be a platform-common feature.
Note that this mostly moves and renames functions. It does not change/modify functionality.

Test Plan:
1. fw-util mb --version and ensure VR version is displayed accurately
2. Ensure that VR sensors are displayed.

Reviewed By: sdasari73

fbshipit-source-id: 1803ba4