Allow ipv4 address printed on USB debug card info page
Summary:
Check if we have a valid IP address (v4 or 6) before printing to the USB debug card.
Thus if the BMC has just a v6 address, we do not waste space in the LCD screen to print the V4 address.

Test Plan: test on FBTP and connect a LCD debug card and ensure v4 address is printed.

Reviewed By: sdasari73

fbshipit-source-id: 34fbd74