openbmc: Always use correct release machine name in /etc/issue
Summary:
If the latest git tag matches the currently built machine then use it in
/etc/issue else form the version from the head commit hash.
For example, if we are building for fbtp, and the latest git tag is 'fbtp-v2.0-<blah>'
then use it as the version in /etc/issue. Else if the latest git tag is
'fby2-v1.2-<blah>' then form the version out of the machine name and the head commit
hash (Example: fbtp-b059382).

Test Plan:
1. Ensure that the latest tag is fbtp-v2.0 (+a few commits on top of that) and do a clean
   build of fbtp (with poky/.sstate deleted).
   Check generated /etc/issue contains: OpenBMC Release fbtp-v2.0-2-gf79c97e-dirty
2. Use 'git tag -a something -m 'something'' to create a random head tag not matching the
   current machine (fbtp) we are building for. Do a clean build (with poky/.sstate deleted)
   Check generated /etc/issue contains: OpenBMC Release fbtp-b059382

Reviewed By: benwei13

fbshipit-source-id: a0ba7ee