vboot: Add PKCS11 support to signing scripts
Summary:
This is the first commit of several that adds hardware-token support for vboot.
In this commit we update the README for testing to demonstrate how to provision keys, collect their public keys, and sign with a Nitrokey Pro.

The testing signing scripts are updated to use `OPENSSL_ENGINES` and `PKCS11_MODULE`.
These environment variables and the new mutex-CLI args for scripts that sign will allow `mkimage` and OpenSSL to use a PKCS11 device.

Test Plan:
1. Run the `create-tests.sh` script, without having anything configured for hardware-tokens.
2. Follow the guide in the README with a Nitrokey Pro.
3. Wait for updates to the `mkimage`, cherry-picked changes from upstream.

Reviewed By: sdasari73

fbshipit-source-id: 944138e