power-util: Add Power Policy Check for 12V Power Control
Summary:
Modify PWR_SLED_CYCLE log string.
Remove set LPS to off in 12V-off.
Add Power Policy and LPS Check for 12V-cycel and 12V-on.
Add enum in obmc-pal.h for power restore policy state.
Make pal.c pal_get_chassis_status to call power restore policy state enum.

Test Plan:
Build and Test with BC DVT System.
Tested PP is off and LPS is any, 12V-cycel and 12V-on will not power on Server.
Tested PP is on and LPS is any, 12V-cycel and 12V-on will power on Server.
Tested PP is lps and LPS is off, 12V-cycel and 12V-on will not power on Server.
Tested PP is lps and LPS is on, 12V-cycel and 12V-on will power on Server.

Reviewed By: jmehta

fbshipit-source-id: 3b6373a