fby2: Add prompt in hsvc-util for hotservice action
Summary: 1. Add prompt in hsvc-util to inform that the SLED is ready to be pulled out for hotservice

Test Plan:
Build and Test pass with EVT System.

root@bmc:~# hsvc-util slot4 --start
Delay 30s for graceful-shutdown
Shutting down fru 4 gracefully...
12V Powering fru 4 to OFF state...
SLED is ready to be pulled out for hot-service now

Reviewed By: amithash

fbshipit-source-id: 671d0b1