fbtp: Remove threshold for local inlet sensor
Summary:
The local inlet sensor may hit UCR (40C) when tested in 35C ambient.
Since we can read both local and remote inlet sensors it is not needed to set UCR temperature for the local inlet sensor.
Hence the proposal to cancel the UCR for the local sensor and base thresholds on the remote inlet sensor only.

Test Plan: Load on FBTP and ensure no regression is observed.

Reviewed By: benwei13

fbshipit-source-id: 0242ff9