Remove package management image feature
Summary:
package-management, aka smart, aka smartpm, aka python-smartpm is a python2 application
which tries to act like yum or apt. We aren't currently using it, but we are building and
including it in our default image.

There was an upstream fork, circa 2012, that converted this project to py3, but it hasn't
been committed to since 2012 and based on an open git issue since that time, the upstream
developers have no interest in migrating their codebase to py3. So for us, it's either
convert and start maintaining this project, or drop it.

I've spoken to tfang about this and we've decided to remove it for now.

Since most of our platforms don't have enough space for py2 and py3, we can't completely
remove py2 without also removing package-management. Hence, let's not use this for our
low-flash-memory systems. We can consider re-adding this for platforms such as CMM, which
do have enough flash space to include both py2 and py3.

Test Plan:
Removed and built wedge100. 'smart' and its dependencies no longer in pn-depends.dot
Image itself no longer has 'smart' python script.

Reviewed By: vineelasmile

fbshipit-source-id: f2d6844