Prepend timestamp to console log files.
Summary:
Modify server to add timestamps to lines read from the
tty. An output is considered a line if newline character is
encountered, even if there are NULL characters on the way. This
allows logging valid string lines as well as (potentially) useful
characters.

Format of log timestamp:

    Fri Sep  8 11:59:28 2017 0000000 message 1
    Fri Sep  8 11:59:28 2017 0000001 message 2

An increasing integer is added so that order can be easily reconstructed.

Test Plan:
tested on wedge hw; run server and client make sure logs are
properly prefixed with timestamps

Reviewed By: tfangit

fbshipit-source-id: ff158b0