Dependency addition of update-rc.d-native to bitbake files which install daemons
Summary:
Looks like this is unnecessary under krogoth builds, but with Pyro, there seems
to be some enhancements to isolate the builds of individual components. So, missing
DEPENDS would ensure a build failure even if the component happened to be built/available
before the particular component is being built.
Also we need to have update-rc.d-native to be added to depends if the bb file has an
explicit install script which calls update-rc.d.

Test Plan: Build krogoth/Poky and ensure Pyro/Poky build goes past the offending failure.

Reviewed By: benwei13

fbshipit-source-id: 05e1ec7