fscd2: Set the value to outmin, only when the value is bigger than outmin
Summary:
Compare the input PWM value with outmin first.
Set the value to outmin, only when the value is bigger than outmin

Test Plan:
Build and Test with System Pass.
Using fake sensor cache reading to trigger transitional and boots mode event, and check the outmin is the bigger one.

Reviewed By: jmehta

fbshipit-source-id: 452fbf6