fby2: Remove redundant check which are already performed by fw-util
Summary:
Remove setting the timeout for firmware update since a much larger value is being set
by fw-util (10min). Considering we now set it back to 0 after the update process this
value no longer need to be accurate (Since it is now used to recover if someone kills
fw-util).
Also remove the deletion of the lock file from bic. It is taken care of by fw-util.

Test Plan: Build on fby2.

Reviewed By: benwei13

fbshipit-source-id: 31d9370