Add support for rest-api testing - diff 1/4
Summary:
This diff adds harness to testing restapi. It is generic enough
to work for network/server/compute.
It takes a json as input that maps endpoint -> "expected data" this expected data
is currently the "keys" in the restapi query. I belive this should prevent issues we
have seen in the last 2-3 helium releases on network side to catch issues much before
than in canary phases.

Also, can be very well leveraged while adopting to asyncio on server/compute restapi

Test Plan:
manual run python cli and flitr
https://our.intern.facebook.com/intern/flitr-jobs/02f34aa23
--> restapi fails because it ran on fido based image and is working
as expected

Reviewed By: tfangit

fbshipit-source-id: 6c0d8b0