Fixup build errors when the build dir is outside the OE root.
Summary:
Building of u-boot and base-files packages fails
when the build directory is outside the OE root.

bitbake fails with the following error:
| DEBUG: Executing shell function do_install | fatal: Not a git repository: ''

The problem is the same for both the packages where do_install scripts
erroneously assume that the working dir is inside the OE root.

This patch fixes the scripts.