fw-util: Remove debug print when alias is added to unknown component.
Summary:
It is normally OK to add aliases to all components we expect on a platform.
Hence it is possible for certain configuration to not contain a physical component thus making an alias invalid.
This change converts the warning print into a debug print since it is expected at normal operation.

Test Plan: Build fbtp.

Reviewed By: benwei13

fbshipit-source-id: 78e7d29