Cache the hostname to avoid repeated lookups on each reboot
Summary:
If the hostname is not set by the name server (most dev environments)
then each nslookup can take up to 10-40 seconds. This is multiplied by 5 (retries).
Hence, optimizing the boot-up-to-ssh time for the failure case by limiting the lookup to the first boot after which hostname is cached in persistent store.
This now causes another set of issues if the hostname changes after the system is provisioned (Causes it to be permanently set to the first hostname).
Hence the second stage where we start a background task which gets the name from ns and checks if the name has changed from the cached copy.
If it has, we will change the cache along with the hostname. syslog is also started so it reflects the new hostname.

Test Plan:
1. time /etc/init.d/hostname.sh twice on both lab and DC machines. The first is expected to be longer than the second (significantly longer for the lab machine).

2. Live testing on both lab and DC machine:
2a. Remove /mnt/data/hostname, reboot. Check `hostname` after the machine comes up. It should be what we expect. The lab machine should have a period where it is ping-able but not ssh-able (hostname blocking the init process).
2b. Reboot (leaving /mnt/data/hostname). Check `hostname` after the machine comes up. It should be what we expect. The lab machine should have a much shorter (if any) period where it is ping-able but not ssh-able.
2c. Manually edit /mnt/data/hostname to change it to something else. Reboot and ssh as soon as possible and execute `hostname`.
      We should see the edited "wrong" hostname (ps | grep nslookup should show it running for some IP address). After a while, `hostname` should show the correct expected hostname.
       logger "Hello world" and check /var/log/messages. Earlier messages should have the "wrong" hostname while the "hello world" one should have the correct hostname.
       PS: this might not be easy to test on DC since the nslookup portion happens within a second. But we should be able to see the change in /var/log/messages.

fbshipit-source-id: b9e472f