Fix potential stack overflow in forming ipmb path
Summary:
gcc7 complained that the path formed "%s_%d" could exceed the 20 characters.
(16 for %s, 1 for '_' and the %d considering a uint8_t could be up to 3 characters
 and finally the nul terminating character - thus bringing the usage to 21 chars).
This is less of a possibility since the bus numbers rarely go to double digits leave
alone single digits. But fix it since our assumptions could change in the future.

Test Plan: build fbtp

Reviewed By: benwei13

fbshipit-source-id: d6c41c6