Fix stack overflow in forming SEL
Summary:
sprintf is used to print "%02X%02X%02X" onto event_data later on in the function.
This clearly is 6 characters (+1 for nul terminating character). Thus, causing
a buffer overflow on the event_data char array. Increasing this to 8 bytes.
PS: This was a compiler warning from gcc7.

Test Plan: build on fbtp

Reviewed By: benwei13

fbshipit-source-id: 682af6a