fbtp: fix the retry in reading INA230 sensors
Summary:
Summary
-If the system is powered off in the period of reading the C2_P12V_INA230 sensors, BMC will get the abnormal reading since the power lost.
But the reading will be keep in the power off state since the static variable is used.
When the system is powered on, BMC will use the last reading and return it to sensord because of an unreliable judgment.
Therefore, user can see the lower critical is asserted and it will be deasserted in the next loop.

-Remove the static variable to make sure that the reading is real in each monitor loop

TestPlan
-Build code(pass)
-Run the DC cycle 3 days and no errors were found(pass)
-Run the mprime 3 days and no errors were found(pass)

Reviewed By: benwei13

fbshipit-source-id: be0ee1f