u-boot: Edit the config file under the build staging directory
Summary:
In Krogoth, both the source (${S}) and the build staging directory (${B}) are the same or
at least the edits happen before it gets copied into the staging directory.
In Rocko, this is a different directory so the edits performed do not get used by the
actual u-boot builds thus causing very weird issues relating to verified boot.

Test Plan:
1. Build tioga pass using github variant with rocko version of yocto (Pass)
2. sign the generate image and load on both ROM and BMC of FBTP. Reboot. The image should come up with all partitions visible in /proc/mtd and no immediate issues observed.
3. Repeat (1) and (2) for krogoth yocto version (Current stable) and ensure no breakage.

Reviewed By: sdasari73

fbshipit-source-id: fac9952