sensor-correction: Limit scope of sensor list to the library
Summary:
Limit scope of variables to the source file. They do not have to
be visible externally. This can pollute global namespace and especially
considering this is a shared library can cause serious issues when
similar names are used by other libraries and linking process.

Test Plan: Load on fbtp and ensure no regressions are observed.

Reviewed By: sdasari73

fbshipit-source-id: 5f69348