Avoid some do_install tasks to crash when using git submodules
When meta-openbmc is clone as a git submodule, the git directory is contained in
the parent project .git/modules directory. In this case, meta-openbmc/.git is a
file containing the git directory location rather than a directory, causing some
do_install scripts to fail.

This patch check that meta-openbmc/.git exist without assuming it is a directory.