tpm-tools: Add patch to fix warnings in new compilers
Summary:
Newer GCC warns if there is indentation which can be confusion for example:
if (condition)
something()
It assumes that the user did not intend something() to be part of the if block and throws a warning.
Unfortunately, tpm-tools marks warnings as errors and the build fails.
This adds a patch which fixes this.

Test Plan: Build under krogoth and rocko

Reviewed By: jmehta

fbshipit-source-id: 3116d99