openbmc-utils: able to keep direction during gpio_get
Summary:
gpio_get() internally changes the GPIO direction to input. This is not
desired in some cases when we just need to read the value when the
direction is still out. Add a 2nd parameter to this function to be able
to 'keepdirection' when reading GPIO value.

Test Plan:
run 'gpio_get SUP_PWR_ON keepdirection' and then check the
'direction' sysfs file to make sure direction is not changed from 'out'
to 'in'.

Reviewed By: mikechoifb

fbshipit-source-id: a783ac6