AST: introduce libcpldupdate-dll-ast-jtag
Summary:
libcpldupdate is a framework to provide hardware control for FPGA/CPLD
upgrade. libcpldupdate-dll-ast-jtag follows the framework and support
the JTAG pin control through sysfs. It relies on the SW control mode of
the JTAG controller in Aspeed BMC SoC.

Test Plan: tested as part of other diff using this library

Reviewed By: mikechoifb

fbshipit-source-id: 7d977b7