fby2: add syslog message for open SLOTX_EJECTOR_LATCH while SLED is seated
Summary:
SLOTX_EJECTOR_LATCH is used during hot service flow to detect if a particular
host is about to be removed.
If an open SLOTX_EJECTOR_LATCH is detected while SLED is seated, it's likely
a faulty latch that needs to be examined and repaired. Add different syslog
message for this condition so we may detect this event in the fleet

Test Plan: verified on YosemiteV2 PVT system

Reviewed By: mikechoifb

fbshipit-source-id: 640ab14