cmm: apply "ignore_probe" param to pfe3000 driver
Summary:
"pfe3000" driver may fail to read device manufacture model in probe()
function; this is a known issue and READ_MFR_MODEL is just commented
out in kernel 4.1 pfe3000 driver.

The root cause is not clear yet (might be caused by flaws in either
hardware or driver, or both), but let's just ignore the failure and
continue device binding for now.

Test Plan:
1) booted up cmm-4.17 and verified all 4 pfe3000 devices are claimed
   by pfe3000 driver successfully.

Reviewed By: covracer

fbshipit-source-id: 6b5c8058e