fbtp: Decide on the machine configuration based on the card type
Summary:
Enhance the machine configuration detection logic to key off
of PCIE card type to determine the machine type. This will
help in differentiating between an AVA card in slot1 vs a
retimer card in slot1 and we could potentially use different
thermal configuration for these systems.

An additional enhancement to add a key which provides the
full description of the machine type to make it easy to
debug (Not for functional use).

Test Plan:
Tested on all platform types and configurations to ensure that
we are getting the correct machine in /mnt/data/kv_store/mb_system_conf.

Thermals are also tested on both platforms

Reviewed By: covracer

fbshipit-source-id: 7c2595ff1