vboot: Add a verified boot support API to libvbs
Summary:
Add support for clients to determine if verified boot
is supported by the current system.
Go one step ahead of determining if the platform has a valid vboot partition
by verifying the verified boot status structure is valid (which is populated by
the SPL).

Test Plan:
1. Build fbtp and execute vboot-util to make sure it shows status as required.
2. Build yosemite (platform not supporting verified boot) and load on system:
root@bmc-oob:~# vboot-util
Verified boot is not supported on this platform!

Reviewed By: benwei13

fbshipit-source-id: 305720926