obmc-pal: Better defaults to the PAL interface
Summary:
Some things dont need to be overridden. Use better
defaults so a platform can choose not to implement
them.

Note this is not the end, but a start in pruning the
PAL API.

Test Plan:
Remove fbtp's pal_get_platform_name() implementation
Then check curl http://localhost:8080/api to ensure that
fbtp is returned.

Reviewed By: vineelasmile

fbshipit-source-id: d4640fc31