Rest endpoint changes to expose the failures
Summary:
While debugging nightly failures for CMM , observed that
trunk is broken for sensors and restapi due to some missing sensors.
Trunk is moved to 4.18 for CMM and is work in progress.
This diff add more verbose logging to restapi test for debuggability.
Additionally adds CMM nightly json for testing.

Test Plan: Details in T37777380

Reviewed By: benwei13

fbshipit-source-id: 4b8768792