fbtp: Upgrade tiogapass to rocko
Summary: Upgrade tiogapass to Rocko.

Test Plan:
FBTP has been testing under rocko for a while and no issues are observed.
One observed issue is that it is easier to observe
I2C bus hang on bus# 6 because of TPM. Currently it has been
reproduced on Krogoth version of fbtp-v6.0 (Even though a bit more rare)

Reviewed By: tomrepo

fbshipit-source-id: 427342e95