cmm: remove "devmem" write operations from setup_gpio.sh
Summary:
Given pinctrl is already handled in device tree in kernel 4.17 (or higher
versions), it's not necessary to update scu registers for pinctrl in
"setup_gpio.sh".

Instead of removing "devmem" command, the command is used to verify if
gpio pins are configured properly in "setup_gpio.sh".

Test Plan:
1) "bitbake cmm-image" succeeded.
2) booted up cmm-4.18 and verified all the gpio pins are exported
   successfully.

Reviewed By: benwei13

fbshipit-source-id: 33c61976d