libgpio: rename functions to avoid symbol resolution conflicts
Summary:
Given "libgpio" and "libgpio-ctrl" may co-exist in openbmc, we need to
make sure the two libraries export different symbols: this is to avoid
symbol resolution conflicts if some applications are linked with both
libraries directly/indirectly. For example:
    App-A --> libX -> libgpio
    App-A --> libY -> libgpio-ctrl

Note: this is a "workaround" during migration, and our ultimate goal
is to move all the applications to the new "gpio-ctrl" library.

Test Plan:
1) openbmc-ci passed.
2) booted up minipack and everything looks normal.

Reviewed By: amithash

fbshipit-source-id: 6ba045d93