BMC reboot will not trigger TH3 reset
Summary:
[YAMP] BMC reboot will not trigger TH3 reset
We have previous check-in D13445414. This is a workaroudn from the vendor to deal with intermittent TH3 ASIC disappearing issue on newer SUPs of YAMP.
But this workaround now introduced a new toggle reset of TH3, whenever BMC is rebooted.
This basically made BMC upgrade or reboot as a disruptive process.

This diff fixes this issue by turning on the board (and toggle reset TH3) only when it is required.
Therefore rebooting BMC or upgrading BMC will not be disruptive any more.

Test Plan: P60889268

Reviewed By: tomrepo

fbshipit-source-id: 3b8de0bb5c