Add TravisCI configuration and update build badges
Summary:
This adds a `.travis.yml` configuration to replace the Jenkins cont-test jobs.
Follow up work is needed to configure TravisCI to run for this GitHub project.

There are some KPs with this approach.
- TravisCI builds are limited to 2CPU and 50 minutes of execution time.
- The build matrix (one build per machine type) does not expose per-machine build badges.

The benefits include easy new machine addition, simply edit the yaml file.

Test Plan: Here is an example of me running this diff on a fork: https://travis-ci.com/theopolis/openbmc-test/branches

Reviewed By: vineelasmile

fbshipit-source-id: df446185be