yosemite: increase healthd mem reboot threshold to 95%
Summary:
as title, increase mem threshold that triggers reboot.
Yv1 had this set to 80%, which is too conservative (as free mem does not
take into reclaimable memory into account). Set this to 95% to be consistent
with other other openBMC platforms

Test Plan:
manually update the json config on test system and verified reboot
is triggered

Reviewed By: amithash

fbshipit-source-id: d16d3ed4ae