fbtp: Recover from hanging mutex held by sensord
Summary:
Currently if a process holds the riser mux mutex in shared
memory and crashes, we end up with a deadlock with reboot
being the only recoverable mechanism.

This change converts the mutex into a "robust" mutex which
will allow us to recover from a crashed process. Alternatives
from research is to use flock() but has the disadvantage of
any fd close to cause it to unlock. My alternative was sem*
APIs but was shocked that they are not recovered if a process
dies with active semaphore gets.

Test Plan:
Repeat in loop restarting sensord and ensure it passes each time to get the correct riser sensors
```
dotest() {
  sv stop sensord > /dev/null
  rm -f /tmp/cache_store/riser_slot2_sensor* # Remove cached values
  sv start sensord > /dev/null
  sleep 10 # Wait for the values to be read
  vals=$(sensor-util riser_slot2 | grep " NA " | grep -v "MB_C2_NVME_CTEMP\|MB_C2_2_NVME_CTEMP\|MB_C2_3_NVME_CTEMP")
  if [ "$vals" != "" ]; then
    echo "FAILED: $vals"
  else
    echo "PASSED"
  fi
}
```

Reviewed By: benwei13

fbshipit-source-id: 858108ff61