Add delay for fan read/set
Summary:
We came accross an issue where all the citv2 test for fan speed
fails, but sometimes they succeeded too. After putting lots of debugging
log, I found out they were failing because it couldn't access the sysfs
path and an exception was being thrown because it doesn't see that the
path exist. I wrote a seperate small script to test access the sysfs
path by itself. The script was succeeded even after 20 seperate run. At
this point, I become suspicious that the issue might not be related to
the i2c sysfs path, but instead due to some delay that is needed in the
fan test. We will run some stress test ont this to see what the outcome
is with all the delay that I added.

Test Plan: Please see this paste P67944745

Reviewed By: mikechoifb

fbshipit-source-id: f0acc9b76e