openbmc-fb: Set DISTRO_FEATURES correctly
The variable $DISTRO_FEATURES_LIBC is no longer used in
OE-Core, this causes errors during boot since the system
was expecting ipv4 by default (busyboxs config changes
depending on features, and it affects things like ifup).

Add ipv4 to OPENBMC_FB_DEFAULT_FEATURE to fix the issue,
and only keep DISTRO_FEATURES_LIBC for compatibility,
but don't rely on it since it is not being set on
default-distrovars.inc on upstream anymore.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>