Adding logic for fpga upgrade issue
Summary:
We recently found in yamp that fpga upgrade depends on jtag
driver which is needed by a shared library file. The path that this
driver is using under kernel 4.1.51 is different from what kernel 5.0 is
using. Therefore, we had to add the logic to take care of that so that
fpga upgrade in 5.0 kernel can succeed.

Test Plan:
Please see paste P79254632 which shows fpga upgrade on kernel
5.0

Reviewed By: tomrepo

fbshipit-source-id: 559405a33d