cit2: cmm: remove a few kernel modules from checklist
Summary:
Generally speaking, these modules are removed from checklist because
they don't have to be compiled as modules. For example, we may decide to
compile them into kernel for some reasons in the future.

If we want to test the functionality provided by these modules, then we
should do that in different places. For example, i2c-mux-pca954x needs
to be covered in i2c test (and we've already done that).

Notes: I'm suspecting "crypto_hash" is automatically enabled with ext4
but I'm not quite sure. So let's keep it in the checklist as of now.

Test Plan: cmm CIT passed.

Reviewed By: benwei13

fbshipit-source-id: 9f59f83f64