Add updated nlohmann JSON package
Summary:
Remove sources of nlohmann JSON from our repo and point to source.
This makes upgrading a lot easier since we do not directly contribute to that project, there is no reason to duplicate that in openBMC.
Upgrade to version 3.3 which brings cmake support as well.

Test Plan: Build fbtp and libadc which uses nlohmann-json

Reviewed By: benwei13

fbshipit-source-id: 6676d78076