fbnd: Block power control when a crashdump is in progress. (#1001)
Summary:
1. ND crash-dump is two part:
(1) The first is BIOS send MCA bank information to BMC and BMC save it in a temp file.
(2) The second part is running crashdump_nd.sh to collect all other information and package it.
The original first part was not protected, so increase the protection of this part.
2. modify crash-dump tmp file Save to /tmp.
Pull Request resolved: https://github.com/facebookexternal/openbmc.quanta/pull/1001

Test Plan:
1.Build and test pass on fbnd system.

root@bmc-oob:/var/run# power-util slot1 cycle
Crashdump for slot1 is ongoing, block the power controlling.
ERROR: power-util fru[1] [POWER_CYCLE] failed

Reviewed By: zhdaniel12

fbshipit-source-id: 195c8eb09c