openbmc-fb.conf: include realpath in HOSTTOOLS
Summary:
Add "realpath" in HOSTTOOLS so linux kernel 5.2 (or higher versions) can
be baked properly. Without the patch, "scripts/mkmakefile" would fail
with "realpath not found".

Note: "realpath" is already added to HOSTTOOLS in the latest yocto releases,
so it's safe to remove the line when we move out of rocko.

Test Plan:
1) bitbake linux-aspeed 4.1 kernel: passed.
2) bitbake linux-aspeed 5.3 kernel: passed.

Reviewed By: amithash

fbshipit-source-id: 0cb39d6a2b