Allow monitoring sensors sourced from the host
Summary:
Due to hardware limitation some sensors might be accessible from the host server
but not from the BMC. An earlier diff added support to allow the server to send
the sensor value via the standard set-sensor command.
This change adds the support for it to be treated as a first class sensor in the BMC
(Allow sensord to monitor its threshold)

Platforms needs to allocate a sensor ID for it and the various thresholds which
tends to be platform specific. Requirement from them is to ensure
1. pal_get_fru_sensor_list should include this sensor,
2. newly added pal_sensor_is_source_host should return true for the sensor.
3. sdr_get_snr_thresh (And in turn: pal_sensor_sdr_init or pal_get_sensor_* should
   support and return valid values for these sensors)

Test Plan: Built together with FBY2 support and tested with a daemon running in server

Reviewed By: benwei13

fbshipit-source-id: e318fcfe6a