Remove support to coexist with py2 based recipes
Summary:
Back in the krogoth days we need to co-exist with some recipes needing
python2 and getting them to support py3. That is changed now.
Everything in openbmc is py3-only and since Rocko, even yocto
recipes have upgraded to using py3.
These changes served a purpose in the krogoth world, but outdated in
the rocko and future world.

Test Plan: Build fbtp, flash on fbtp ensure python2 is not packaged.

Reviewed By: sammyblueeyes

fbshipit-source-id: be8d16d039