test2: interface_test: add more debug data
Summary:
I have seen intermittent testcase failures in test_eth0_v4_interface
complaining about a "list index out of range" exception in this code:
    ipv4 = out.split("inet ")[1].split("/")[0]

My suspicion is that the device does not have an ipv4 address and the
`ip addr` call, therefore, does not have a "inet " string.  (This seems
to be the only place where we could get an index error in this code.)

Add an exception when a check for "inet " and "inet6 " fails that
includes the full `ip addr` output so we can better determine what is
happening.

Test Plan: netcastle run.

Reviewed By: vineelasmile

fbshipit-source-id: d9413be01a