travis-ci: fix env source arguments
Summary:
With D16954252, amithash simplified the arguments required by the
sourcing of openbmc-init-build-env.  A side-effect of this is that
we can support machines where the meta-foo and conf/machine/bar.conf
are different names.  Simplify the way .travis.yml calls the
environment setup to use this new method.

Test Plan: None available.  Travis currently broken.

Reviewed By: sammyblueeyes

fbshipit-source-id: 4530576c0a