travis-ci: remove sstate caching
Per https://docs.travis-ci.com/user/caching/#things-not-to-cache
we should not be caching large compiled binaries, which is effectively
what the sstate-cache is.  Currently we are telling TravisCI to "cache"
this build output directory but then never using it and they are tens
of GB in size.  Disable this caching feature and simply download the
latest sstate cache.

Signed-off-by: Patrick Williams <patrickw3@fb.com>