Ensure we are flashing an image for the correct platform
Summary:
Check that the platform encoded in the U-Boot metadata
matches the current platform. If it doesn't, we abort the flashing
unless we are running with `--force`.

If the image is old and doesn't have metadata or a valid `/etc/issue`
we allow the flashing to continue, to cause minimal disruptions to
automation.

Test Plan:
Tested on a Wedge100. Tried to apply a minipack image with

```
$ /tmp/improve_system.py --dry-run /tmp/flash-wrong
[...]
INFO:2020-01-23 08:37:20,839 Checking the sanity of the request
WARNING:2020-01-23 08:37:29,520 Image /tmp/flash-wrong is not for this platform!
ERROR:2020-01-23 08:37:29,558 Aborting flash. Use a valid image or run with --force
```

But with a correct image, we have:

```
$ /tmp/improve_system.py --dry-run /tmp/flash
[...]
INFO:2020-01-23 08:39:08,117 Checking the sanity of the request
INFO:2020-01-23 08:39:16,243 Proceeding with flash
```

Reviewed By: kawmarco

fbshipit-source-id: 752ec6c04c