kv: set: catch values longer than MAX_VALUE_LEN
Summary:
kv values should only be MAX_VALUE_LEN.  The value requested to be
stored is longer than that return E2BIG.

Also, when no length is provided we were using strlen to calculate
the length of the value (assuming it is a string).  Switch to strnlen
to avoid buffer overruns and return E2BIG for strings which are not
null terminated within (MAX_VALUE_LEN-1) bytes.

Test Plan: Add unit tests to catch both cases.

Reviewed By: zhdaniel12

fbshipit-source-id: e54a70f1e4