kv: get: ensure null terminator is added to strings
Summary:
The kv_get function allows string-type values to be read from the
kv-store by passing a NULL length output variable.  In that case,
the caller has no way to know the length of the value read and likely
uses something like strlen to calculate it.  Ensure there is a null
terminator added to the value in that case and truncate if necessary.
Warn in the syslog in cases of truncation.

Since kv_set does not store a null terminator into the file, this also
helps in cases where the input buffer to kv_get were not initialized
because we ensure a null terminator is still added.

Test Plan: Add unit test to cover.

fbshipit-source-id: 4ad694fcab