openbmc_version: reverse tag search order
Summary:
Recently we made two tags for sonorapass that pointed to the
same commit (v2020.02.0 and 02.1).  When the image for 2020.02.1
was built it was observed that the version string was still 02.0.

`git tag` will list tags in alphanumerical sort, so lower versions
will appear sooner.  This is the opposite of what we want since we
would want newer / higher tags to be used for the image build.

Reverse the list to accomplish this.

Test Plan: Create some fake tags against this commit and confirm expected behavior.

Reviewed By: vineelasmile

fbshipit-source-id: 2db64e233b