fby3: re-initialize the network of BMC to fixup IPv6 issue (#917)
Summary:
-When BMC performs the sled-cycle, IPv6 may be lost next time.
The lost of IPv6 can be recovered when BMC re-initializes the network of BMC.
And so, add this script as a workaround to fixup the issue.
The script will be removed when the issue is solved.
Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/917

Test Plan:
-Build code(pass)
-Check eth0_ipv6_fixup.sh is created in rc5.d(pass)
-Check eth0_ipv6_fixup.sh is run(pass)

Log:
Set up GPIO pins.....Running /etc/init.d/networking restart is deprecated because it may not enable again some interfaces
Reconfiguring network interfaces...

Reviewed By: vineelasmile

fbshipit-source-id: 53c9b5a44a