common: replace ntp-utils with ntpq
Summary:
ntp-utils is a sub-package of ntp and includes user-oriented tools
related to NTP.  Some of these tools are perl scripts and thus cause
us to pull in perl.  Remove this package since it seems to be unused
and unneeded.

We do use ntpq, which is also part of ntp and pulled in as an RDEPENDS
from ntp-utils.  Replace the image dependency of ntp-utils with ntpq
to keep ntpq in the image.

Also, add perl to PACKAGE_EXCLUDE to ensure it is not installed and
fails a build if it is ever installed.  We've had multiple attempts
to remove perl from our install image and it continues to sneak back
in.  Adding it to PACKAGE_EXCLUDE ensures it does not.

Test Plan: Built and booted on Sonora Pass

Reviewed By: zhdaniel12

fbshipit-source-id: 021d836ba4