common: remove ssl CA certificates
Summary:
Nearly every package that uses TLS/SSL RRECOMMENDS the CA certificate
package.  Since we should not be initiating SSL traffic from a BMC
to outside of our own network, there is no value in all of these
certificates.  Add it to the BAD_RECOMMENDATIONS so it is not installed
unless there is a strict need for it.

Test Plan: Built and booted on Sonora Pass

Reviewed By: tfangit

fbshipit-source-id: c2b58e6359