rng-tools: remove libjitterentropy
Summary:
The warrior version of rng-tools introduced an optional PACKAGECONFIG,
which was enabled by default, called `libjitterentropy`.  Since the
BMC chips have a hardware RNG, this is not a necessary feature, and
causes a noticeable delay in the start of the rngd.  On systems with
very low IO activity we are seeing cases of what appears to be hangs.

Disable the PACKAGECONFIG.

Test Plan: Built and booted on Sonora Pass; observed faster daemon startup.

Reviewed By: benwei13

fbshipit-source-id: 76b794f123