python3: add fcntl package
Summary:
Havoc deploys a python script onto the BMC which has an `import fcntl`,
which worked fine on yocto-rocko systems, but does not work with
yocto-warrior.  Insert a dependency in the python package group on
python3-fcntl so the module is available.

Test Plan: Build sonorapass and check package manifest.

Reviewed By: benwei13

fbshipit-source-id: d7d541bef6