obmc-pal: create meson built v0.2
Summary:
The refactoring of obmc-pal to be build using meson wasn't able to
be done in a backwards compatible way.  Define a new version of the
package (v0.2) that can be built with meson, but restrict all platforms
to defaulting to the current v0.1.

Test Plan: Build and boot on sonorapass.  Pass CIT.

Reviewed By: sammyblueeyes

fbshipit-source-id: ae0303c3f8