obmc-pal: v0.1: add fake versioned library symlink
Summary:
Some python libraries use CDLL to directly open the libpal.so.  By
moving to meson and building versioned shared libraries, libpal.so
will no longer exist but instead libpal.so.0 will exist.

In order to allow the python scripts to work with either version of
the libpal package, we are going to create a fake libpal.so.0 symlink
for v0.1 and a real libpal.so.0 for v0.2, and migrate the python
scripts to load libpal.so.0 in all cases.

Test Plan: Build and boot on sonorapass.  Pass CIT.

Reviewed By: sammyblueeyes

fbshipit-source-id: 4c673f8432