pfr-util: Add multi-node support
Summary:
This change refactors the code to isolate the final SMBUS communication into a single function pfr_transfer() thus allowing us to specialize it to be able to send it over BIC to a remote PFR which we do not have direct access via SMBUS.

This also changes the interface to allow user to specify the FRU thus allowing user to specify which PFR (slot1-4 or baseboard).

Test Plan:
Considering this is a code refactor, a build test is sufficient.
This needs to be tested on real hardware.

Reviewed By: benwei13

fbshipit-source-id: 17f4fa4677